Revert project renaming and restructuring
