--- backend/WebApi/Helpers/DBHelper.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace WebApi.Helpers
-{
-    public class DBHelper
-    {
-        public static void PopulateDatabaseFromJson()
-        {
-
-        }
-    }
-}
