+++ backend/WebApi/Startup.cs
@@ -5,10 +5,9 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
-using Persistence.File;
-using Persistence.File.Interfaces;
-using Persistence.File.Repositories;
-using Persistence.EF;
+using Persistence;
+using Persistence.Interfaces;
+using Persistence.Repositories;
 using Services.Interfaces;
 using Services.Options;
 using Services.Services;
